Migrate transpose and matrix correctlyRounded tests
