+++ src/webgpu/shader/execution/expression/call/builtin/transpose.spec.ts
@@ -9,10 +9,10 @@ Returns the transpose of e.
 import { makeTestGroup } from '../../../../../../common/framework/test_group.js';
 import { GPUTest } from '../../../../../gpu_test.js';
 import { TypeF32, TypeMat } from '../../../../../util/conversion.js';
-import { transposeInterval } from '../../../../../util/f32_interval.js';
+import { FP } from '../../../../../util/floating_point.js';
 import { sparseMatrixF32Range } from '../../../../../util/math.js';
 import { makeCaseCache } from '../../case_cache.js';
-import { allInputSources, generateMatrixToMatrixCases, run } from '../../expression.js';
+import { allInputSources, run } from '../../expression.js';
 
 import { builtin } from './builtin.js';
 
@@ -20,58 +20,130 @@ export const g = makeTestGroup(GPUTest);
 
 export const d = makeCaseCache('transpose', {
   f32_mat2x2_const: () => {
-    return generateMatrixToMatrixCases(sparseMatrixF32Range(2, 2), 'finite', transposeInterval);
+    return FP.f32.generateMatrixToMatrixCases(
+      sparseMatrixF32Range(2, 2),
+      'finite',
+      FP.f32.transposeInterval
+    );
   },
   f32_mat2x2_non_const: () => {
-    return generateMatrixToMatrixCases(sparseMatrixF32Range(2, 2), 'unfiltered', transposeInterval);
+    return FP.f32.generateMatrixToMatrixCases(
+      sparseMatrixF32Range(2, 2),
+      'unfiltered',
+      FP.f32.transposeInterval
+    );
   },
   f32_mat2x3_const: () => {
-    return generateMatrixToMatrixCases(sparseMatrixF32Range(2, 3), 'finite', transposeInterval);
+    return FP.f32.generateMatrixToMatrixCases(
+      sparseMatrixF32Range(2, 3),
+      'finite',
+      FP.f32.transposeInterval
+    );
   },
   f32_mat2x3_non_const: () => {
-    return generateMatrixToMatrixCases(sparseMatrixF32Range(2, 3), 'unfiltered', transposeInterval);
+    return FP.f32.generateMatrixToMatrixCases(
+      sparseMatrixF32Range(2, 3),
+      'unfiltered',
+      FP.f32.transposeInterval
+    );
   },
   f32_mat2x4_const: () => {
-    return generateMatrixToMatrixCases(sparseMatrixF32Range(2, 4), 'finite', transposeInterval);
+    return FP.f32.generateMatrixToMatrixCases(
+      sparseMatrixF32Range(2, 4),
+      'finite',
+      FP.f32.transposeInterval
+    );
   },
   f32_mat2x4_non_const: () => {
-    return generateMatrixToMatrixCases(sparseMatrixF32Range(2, 4), 'unfiltered', transposeInterval);
+    return FP.f32.generateMatrixToMatrixCases(
+      sparseMatrixF32Range(2, 4),
+      'unfiltered',
+      FP.f32.transposeInterval
+    );
   },
   f32_mat3x2_const: () => {
-    return generateMatrixToMatrixCases(sparseMatrixF32Range(3, 2), 'finite', transposeInterval);
+    return FP.f32.generateMatrixToMatrixCases(
+      sparseMatrixF32Range(3, 2),
+      'finite',
+      FP.f32.transposeInterval
+    );
   },
   f32_mat3x2_non_const: () => {
-    return generateMatrixToMatrixCases(sparseMatrixF32Range(3, 2), 'unfiltered', transposeInterval);
+    return FP.f32.generateMatrixToMatrixCases(
+      sparseMatrixF32Range(3, 2),
+      'unfiltered',
+      FP.f32.transposeInterval
+    );
   },
   f32_mat3x3_const: () => {
-    return generateMatrixToMatrixCases(sparseMatrixF32Range(3, 3), 'finite', transposeInterval);
+    return FP.f32.generateMatrixToMatrixCases(
+      sparseMatrixF32Range(3, 3),
+      'finite',
+      FP.f32.transposeInterval
+    );
   },
   f32_mat3x3_non_const: () => {
-    return generateMatrixToMatrixCases(sparseMatrixF32Range(3, 3), 'unfiltered', transposeInterval);
+    return FP.f32.generateMatrixToMatrixCases(
+      sparseMatrixF32Range(3, 3),
+      'unfiltered',
+      FP.f32.transposeInterval
+    );
   },
   f32_mat3x4_const: () => {
-    return generateMatrixToMatrixCases(sparseMatrixF32Range(3, 4), 'finite', transposeInterval);
+    return FP.f32.generateMatrixToMatrixCases(
+      sparseMatrixF32Range(3, 4),
+      'finite',
+      FP.f32.transposeInterval
+    );
   },
   f32_mat3x4_non_const: () => {
-    return generateMatrixToMatrixCases(sparseMatrixF32Range(3, 4), 'unfiltered', transposeInterval);
+    return FP.f32.generateMatrixToMatrixCases(
+      sparseMatrixF32Range(3, 4),
+      'unfiltered',
+      FP.f32.transposeInterval
+    );
   },
   f32_mat4x2_const: () => {
-    return generateMatrixToMatrixCases(sparseMatrixF32Range(4, 2), 'finite', transposeInterval);
+    return FP.f32.generateMatrixToMatrixCases(
+      sparseMatrixF32Range(4, 2),
+      'finite',
+      FP.f32.transposeInterval
+    );
   },
   f32_mat4x2_non_const: () => {
-    return generateMatrixToMatrixCases(sparseMatrixF32Range(4, 2), 'unfiltered', transposeInterval);
+    return FP.f32.generateMatrixToMatrixCases(
+      sparseMatrixF32Range(4, 2),
+      'unfiltered',
+      FP.f32.transposeInterval
+    );
   },
   f32_mat4x3_const: () => {
-    return generateMatrixToMatrixCases(sparseMatrixF32Range(4, 3), 'finite', transposeInterval);
+    return FP.f32.generateMatrixToMatrixCases(
+      sparseMatrixF32Range(4, 3),
+      'finite',
+      FP.f32.transposeInterval
+    );
   },
   f32_mat4x3_non_const: () => {
-    return generateMatrixToMatrixCases(sparseMatrixF32Range(4, 3), 'unfiltered', transposeInterval);
+    return FP.f32.generateMatrixToMatrixCases(
+      sparseMatrixF32Range(4, 3),
+      'unfiltered',
+      FP.f32.transposeInterval
+    );
   },
   f32_mat4x4_const: () => {
-    return generateMatrixToMatrixCases(sparseMatrixF32Range(4, 4), 'finite', transposeInterval);
+    return FP.f32.generateMatrixToMatrixCases(
+      sparseMatrixF32Range(4, 4),
+      'finite',
+      FP.f32.transposeInterval
+    );
   },
   f32_mat4x4_non_const: () => {
-    return generateMatrixToMatrixCases(sparseMatrixF32Range(4, 4), 'unfiltered', transposeInterval);
+    return FP.f32.generateMatrixToMatrixCases(
+      sparseMatrixF32Range(4, 4),
+      'unfiltered',
+      FP.f32.transposeInterval
+    );
   },
 });
