+++ src/webgpu/shader/execution/expression/unary/f32_conversion.spec.ts
@@ -15,7 +15,8 @@ import {
   TypeU32,
   u32,
 } from '../../../../util/conversion.js';
-import { correctlyRoundedInterval, correctlyRoundedMatrix } from '../../../../util/f32_interval.js';
+import { correctlyRoundedInterval } from '../../../../util/f32_interval.js';
+import { FP } from '../../../../util/floating_point.js';
 import {
   fullF32Range,
   fullI32Range,
@@ -23,7 +24,7 @@ import {
   sparseMatrixF32Range,
 } from '../../../../util/math.js';
 import { makeCaseCache } from '../case_cache.js';
-import { allInputSources, generateMatrixToMatrixCases, run, ShaderBuilder } from '../expression.js';
+import { allInputSources, run, ShaderBuilder } from '../expression.js';
 
 import { unary } from './unary.js';
 
@@ -52,129 +53,129 @@ export const d = makeCaseCache('unary/f32_conversion', {
     });
   },
   f32_mat2x2_const: () => {
-    return generateMatrixToMatrixCases(
+    return FP.f32.generateMatrixToMatrixCases(
       sparseMatrixF32Range(2, 2),
       'finite',
-      correctlyRoundedMatrix
+      FP.f32.correctlyRoundedMatrix
     );
   },
   f32_mat2x2_non_const: () => {
-    return generateMatrixToMatrixCases(
+    return FP.f32.generateMatrixToMatrixCases(
       sparseMatrixF32Range(2, 2),
       'unfiltered',
-      correctlyRoundedMatrix
+      FP.f32.correctlyRoundedMatrix
     );
   },
   f32_mat2x3_const: () => {
-    return generateMatrixToMatrixCases(
+    return FP.f32.generateMatrixToMatrixCases(
       sparseMatrixF32Range(2, 3),
       'finite',
-      correctlyRoundedMatrix
+      FP.f32.correctlyRoundedMatrix
     );
   },
   f32_mat2x3_non_const: () => {
-    return generateMatrixToMatrixCases(
+    return FP.f32.generateMatrixToMatrixCases(
       sparseMatrixF32Range(2, 3),
       'unfiltered',
-      correctlyRoundedMatrix
+      FP.f32.correctlyRoundedMatrix
     );
   },
   f32_mat2x4_const: () => {
-    return generateMatrixToMatrixCases(
+    return FP.f32.generateMatrixToMatrixCases(
       sparseMatrixF32Range(2, 4),
       'finite',
-      correctlyRoundedMatrix
+      FP.f32.correctlyRoundedMatrix
     );
   },
   f32_mat2x4_non_const: () => {
-    return generateMatrixToMatrixCases(
+    return FP.f32.generateMatrixToMatrixCases(
       sparseMatrixF32Range(2, 4),
       'unfiltered',
-      correctlyRoundedMatrix
+      FP.f32.correctlyRoundedMatrix
     );
   },
   f32_mat3x2_const: () => {
-    return generateMatrixToMatrixCases(
+    return FP.f32.generateMatrixToMatrixCases(
       sparseMatrixF32Range(3, 2),
       'finite',
-      correctlyRoundedMatrix
+      FP.f32.correctlyRoundedMatrix
     );
   },
   f32_mat3x2_non_const: () => {
-    return generateMatrixToMatrixCases(
+    return FP.f32.generateMatrixToMatrixCases(
       sparseMatrixF32Range(3, 2),
       'unfiltered',
-      correctlyRoundedMatrix
+      FP.f32.correctlyRoundedMatrix
     );
   },
   f32_mat3x3_const: () => {
-    return generateMatrixToMatrixCases(
+    return FP.f32.generateMatrixToMatrixCases(
       sparseMatrixF32Range(3, 3),
       'finite',
-      correctlyRoundedMatrix
+      FP.f32.correctlyRoundedMatrix
     );
   },
   f32_mat3x3_non_const: () => {
-    return generateMatrixToMatrixCases(
+    return FP.f32.generateMatrixToMatrixCases(
       sparseMatrixF32Range(3, 3),
       'unfiltered',
-      correctlyRoundedMatrix
+      FP.f32.correctlyRoundedMatrix
     );
   },
   f32_mat3x4_const: () => {
-    return generateMatrixToMatrixCases(
+    return FP.f32.generateMatrixToMatrixCases(
       sparseMatrixF32Range(3, 4),
       'finite',
-      correctlyRoundedMatrix
+      FP.f32.correctlyRoundedMatrix
     );
   },
   f32_mat3x4_non_const: () => {
-    return generateMatrixToMatrixCases(
+    return FP.f32.generateMatrixToMatrixCases(
       sparseMatrixF32Range(3, 4),
       'unfiltered',
-      correctlyRoundedMatrix
+      FP.f32.correctlyRoundedMatrix
     );
   },
   f32_mat4x2_const: () => {
-    return generateMatrixToMatrixCases(
+    return FP.f32.generateMatrixToMatrixCases(
       sparseMatrixF32Range(4, 2),
       'finite',
-      correctlyRoundedMatrix
+      FP.f32.correctlyRoundedMatrix
     );
   },
   f32_mat4x2_non_const: () => {
-    return generateMatrixToMatrixCases(
+    return FP.f32.generateMatrixToMatrixCases(
       sparseMatrixF32Range(4, 2),
       'unfiltered',
-      correctlyRoundedMatrix
+      FP.f32.correctlyRoundedMatrix
     );
   },
   f32_mat4x3_const: () => {
-    return generateMatrixToMatrixCases(
+    return FP.f32.generateMatrixToMatrixCases(
       sparseMatrixF32Range(4, 3),
       'finite',
-      correctlyRoundedMatrix
+      FP.f32.correctlyRoundedMatrix
     );
   },
   f32_mat4x3_non_const: () => {
-    return generateMatrixToMatrixCases(
+    return FP.f32.generateMatrixToMatrixCases(
       sparseMatrixF32Range(4, 3),
       'unfiltered',
-      correctlyRoundedMatrix
+      FP.f32.correctlyRoundedMatrix
     );
   },
   f32_mat4x4_const: () => {
-    return generateMatrixToMatrixCases(
+    return FP.f32.generateMatrixToMatrixCases(
       sparseMatrixF32Range(4, 4),
       'finite',
-      correctlyRoundedMatrix
+      FP.f32.correctlyRoundedMatrix
     );
   },
   f32_mat4x4_non_const: () => {
-    return generateMatrixToMatrixCases(
+    return FP.f32.generateMatrixToMatrixCases(
       sparseMatrixF32Range(4, 4),
       'unfiltered',
-      correctlyRoundedMatrix
+      FP.f32.correctlyRoundedMatrix
     );
   },
 });
